requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.4
djangorestframework==3.16.0
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==11.3.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
whitenoise==6.9.0
